Fix eups lib->lib64 link fail build breaker
Some part of the xrootd build creates a lib/python2.7 directory.
Later, when eupspkg.cfg.sh tries to create a link lib->lib64,
it trips over the existing lib directory and drops the link in
the wrong place.

Instead, we now link to the contents of lib64 from within the lib
directory, instead of tyring to link the directories themselves.